Split RenameBuild into rename and copy functions
They do different things and should therefore be different functions. Using boolean params to control function behavior is an antipattern. Escpecially since  copy and rename behave different on case-insensitive file systems. I can rename a file to an existing file (e.g. change casing) but copying to a file with different case would override it.

This changed happend in preparation of a fix for Openarl/PathOfBuilding#709